refactor: split functions into composables & use pinia
