+++ index.html
@@ -12,10 +12,9 @@
 
 </head>
 
-<body onload="timer()">
+<body>
   <div id="app"></div>
   <script type="module" src="/src/main.js"></script>
-
 </body>
 
 
@@ -27,22 +26,4 @@
 <script type="text/javascript"
         src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.3.1/js/bootstrap.min.js"></script>
 
-<script>
-console.log('test')
-    function timer() {
-        const now = new Date();
-        const month = '0' + now.getMonth();
-        const date = '0' + now.getDate();
-        const hour = '0' + now.getHours();
-        const min = '0' + now.getMinutes();
-        let day = now.getDay();
-        const day_arr = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
-
-        day = day_arr[day];
-
-        $('#date').text(now.getFullYear() + '.' + month.substr(-2) + '.' + date.substr(-2) + '　' + day + ' ' + hour.substr(-2) + ':' + min.substr(-2));
-        setTimeout(timer, 1000);
-    }
-
-</script>
 </html>
